backend/face-recognition-tools/requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
keras-facenet==0.0.4
psycopg2-binary==2.9.7
requests==2.31.0
pytz==2023.3
Pillow==10.0.1